[cfme-versions.rb] Simplify `.version` logic
Previously, when we were using `CFME` versions numbers (`:cfme_release`)
for the gem version for this repo, the semver for that product looked
like this:

- 5.8.z
- 5.9.z
- 5.10.z
- ...

So a little more logic was required to ensure we got the proper
numerical parts of the version string, and combined it was a stub
"z-stream" for the `CFME::Versions.version` method.

However, now that the versioning is using `miq_semver`, which is
versions pretty straight forwardly:

- 6.y.z
- 7.y.z
- 8.y.z
- ...

The logic can be simplified, and the Regexp that existed as part of the
getting the proper parts of the `.split` are no longer needed.

Also, the "new logic" (prior to this commit), does a `+ %w[0 0 0]`
followed by a `.first(3)`... which means the 3rd element in that word
array is pointless, and the instructions in the `README.md` didn't align
with what was required.  This change simplifies what is needed to make a
patch going forward, as well as re-clarifies the instructions in the
`README.md`.